Unmute sound unconditionally when application gains focus
This works around a bug where clicking the "OK" button of the config dialog
after disabling the "Mute Sound in Background" checkbox would result in sound
getting muted permanently. The reason for this is that on some systems,
clicking "OK" will result in the application window losing focus while the
new settings have not been applied yet. The result is that the sound gets
muted, and stays muted since the "Mute in background" setting only gets
applied after that.