Add new cfg override for bundling custom fonts
It's useful to be able to ship a game with its own custom fonts rather
than relying on the OS-installed fonts. To load all fonts from the
"fonts" directory, use:

    [BundledFonts]
    fontPath = fonts

in the game's cfg file. "fontPath" needs to be a path relative to the
interpreter's directory.

Supported font formats are those supported by Qt. Currently that's
TrueType (TTF), TrueType Collections (TTC) and OpenType (OTF.)